requirements.txt
dj-database-url==0.4.2
Django==1.11.3
django-mathfilters==0.4.0
django-password-reset==1.0
djangorestframework==3.6.3
gunicorn==19.7.1
paho-mqtt==1.3.0
psycopg2==2.7.1
pytz==2017.2
simplejson==3.11.1
whitenoise==3.3.0
yahoo-finance==1.4.0
